rotors_gazebo/scripts/requirements.txt
actionlib==1.14.0
angles==1.9.13
blessed==1.20.0
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
casadi==3.7.1
catkin==0.8.10
catkin-pkg==1.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
distro==1.9.0
docutils==0.20.1
dynamic-reconfigure==1.7.3
filelock==3.16.1
fonttools==4.55.0
fsspec==2025.3.0
future==1.0.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gpustat==1.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
interactive-markers==1.12.0
Jinja2==3.1.6
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.7
laser_geometry==1.6.7
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==1.17.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
ordered-set==4.1.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.4
resource_retriever==1.12.7
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslz4==1.17.0
rosmaster==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.6.0
rospy==1.17.0
rqt-image-view==0.4.17
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_console==0.4.10
rqt_shell==0.4.11
rqt_top==0.4.10
rqt_web==0.4.10
rviz==1.14.25
scipy==1.10.1
sensor-msgs==1.13.1
six==1.17.0
smclib==1.8.6
sympy==1.13.3
tf==1.13.2
tf-conversions==1.13.2
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.17.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3==2.2.3
wcwidth==0.2.13
xacro==1.14.19
zipp==3.20.2
